Make ShadowSigningInfo properly support parceling
Currently the state of ShadowSigningInfo is not preserved when a SigningInfo is written to a parcel. At the same time ShadowApplicationPackageManager writes PackageInfo objects to parcels and reads them back before returning them. As a result, signing information returned by ShadowApplicationPackageManager is always empty.

To address that, make ShadowSigningInfo support writing to a parcel and creating from a parcel.

PiperOrigin-RevId: 257654358